// DVG.CRM.XeCung.ClientSide/dvg-crm-xecung-final-client/src/app/models/lead-converted.ts
export class LeadConverted {
    constructor() {}
    Id: number;
    LeadAdminId: number;
    Name: string;
    PhoneNumber: string;
    Email: string;
    City: number;
    Region: number;
    Address: string;
    Organization: number;
    OrganizationName: string;
    Group: number;
    CustomerType: number;
    Note: string;
    Token: string;
}